% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authentication.R
\name{authenticate}
\alias{authenticate}
\title{Authenticate access using EUMETSAT's API}
\usage{
authenticate(consumer_key, consumer_secret)
}
\arguments{
\item{consumer_key}{EUMETSAT Consumer Key}

\item{consumer_secret}{EUMETSAT Consumer Secret}
}
\value{
API Access Token as a String
}
\description{
Consumer Key and Consumer Secret are available to authenticated users through
EUMETSAT's EO Portal: https://api.eumetsat.int/api-key/.
}
\examples{
if(FALSE)
{
  authenticate("consumer_key_goes_here", "consumer_secret_goes_here")
}

}
