% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authentication.R
\name{authenticator}
\alias{authenticator}
\title{Authenticate access using EUMETSAT's API}
\usage{
authenticator(
  consumer_key = NULL,
  consumer_secret = NULL,
  save_creds = TRUE,
  return_header = FALSE
)
}
\arguments{
\item{consumer_key}{EUMETSAT Consumer Key (default NULL; attempts to read from file)}

\item{consumer_secret}{EUMETSAT Consumer Secret (default NULL; attempts to read from file)}

\item{save_creds}{Whether credentials are to be stored in a file for later use (default TRUE)}

\item{return_header}{Returns a httr::add_headers object with access token (default FALSE)}
}
\value{
API Access Token as a String or a httr header object
}
\description{
Consumer Key and Consumer Secret are available to authenticated users through
EUMETSAT's EO Portal: https://api.eumetsat.int/api-key/.
}
\examples{
if(FALSE)
{
  authenticate("consumer_key_goes_here", "consumer_secret_goes_here")
}

}
