% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browse.R
\name{collections}
\alias{collections}
\title{Queries the collections endpoint}
\usage{
collections(collID = NULL, authenticated = TRUE)
}
\arguments{
\item{collID}{Collection ID}

\item{authenticated}{whether to send authenticated API requests (defaults to TRUE)}
}
\value{
A list with information about collection(s) and the number of products available
}
\description{
Queries the appropriate \verb{collections/} endpoint depending on the arguments passed to it.
}
\examples{
collections(authenticated = FALSE)
}
