% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovery.R
\name{discover}
\alias{discover}
\title{Queries the EUMETSAT Discover API}
\usage{
discover(hits_only = FALSE)
}
\arguments{
\item{hits_only}{Return details of the hits only}
}
\value{
If \code{hits_only} is FALSE, the entire response is returned as an R list, otherwise only the details of the hits.
}
\description{
Discover EUMETSAT Data Store Earth Observation data at the collection level.
}
\examples{
discover()
discover(hits_only = TRUE)

}
