% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browse.R
\name{footprint}
\alias{footprint}
\title{Returns the footprint for a GEO product}
\usage{
footprint(
  sensorMode,
  subSatelliteLongitude = 0,
  returnPolygon = TRUE,
  plotPolygon = FALSE,
  authenticated = TRUE
)
}
\arguments{
\item{sensorMode}{Sensor Mode; one of: {ALTHRV, REDSCN, NOMSCN}}

\item{subSatelliteLongitude}{Sub Satellite longitude in degrees (defaults to 0 degrees)}

\item{returnPolygon}{TRUE if a polygon should be returned; FALSE to return the response}

\item{plotPolygon}{if the polygon is to be plotted}

\item{authenticated}{whether to send authenticated API requests (defaults to TRUE)}
}
\value{
If returnPolygon is TRUE then a sf polygon else the response as an R list
}
\description{
GEO products do not have an individual footprint. Instead, this resource represents a GEO footprint for the respective sensor mode and sub satellite longitude.
}
\examples{
footprint("ALTHRV", -50, authenticated = FALSE)
footprint("NOMSCN", returnPolygon = FALSE, plotPolygon = TRUE, authenticated = FALSE)

}
