% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browse.R
\name{browse_collections}
\alias{browse_collections}
\title{Queries the collections endpoint}
\usage{
browse_collections(collID = NULL, prodID = NULL, authenticated = TRUE)
}
\arguments{
\item{collID}{Collection ID}

\item{prodID}{Product ID}

\item{authenticated}{whether to send authenticated API requests (defaults to TRUE)}
}
\value{
A list with information about collection(s) and the number of products available
}
\description{
Queries the appropriate \verb{collections/} endpoint depending on the arguments passed to it.
}
\examples{
browse_collections(authenticated = FALSE)
browse_collections('EO:EUM:DAT:MSG:HRSEVIRI', authenticated = FALSE)
browse_collections('EO:EUM:DAT:MSG:HRSEVIRI', 'MSG3-SEVI-MSG15-0100-NA-20240428162742.422000000Z-NA', authenticated = FALSE)
}
